app/feedback-reader/requirements.txt
python-dotenv==1.1.0
psycopg2-binary==2.9.7